+++ ufufs/ufufs.h
@@ -3,9 +3,9 @@
 typedef unsigned int FileDescriptor;
 
 #define MAGIC_NUM -1534
-int ufufs_mount(const char* filePath); 
+int ufufs_mount(const char *filePath);
 
-FileDescriptor ufufs_open(const char* filename);
+FileDescriptor ufufs_open(const char *filename);
 
 int ufufs_read(FileDescriptor fd, void *buf, size_t count);
 
@@ -13,7 +13,7 @@ int ufufs_write(FileDescriptor fd, void *buf, size_t count);
 
 int ufufs_close(FileDescriptor fd);
 
-int ufufs_seek(FileDescriptor fd, int offset);
+off_t ufufs_seek(FileDescriptor fd, size_t offset);
 
 // typedef struct {
 //   const int MAGIC_N;
